backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
redis
httpx
pytest
pytest-asyncio
python-multipart
email-validator
aiosqlite
eval_type_backport
greenlet
